feat(auth): add basic login flow using Auth0
In the process of debugging the Auth0 integration, I switched to a custom
Express server in the hopes that it would fix the problem. It didn't, but it's
what is working now. A future TODO is to see if we can switch back to the
built-in API routes that Next.js exposes.